+++ lib/karafka/instrumentation/stdout_listener.rb
@@ -23,7 +23,13 @@ module Karafka
           consumer = event[:consumer]
           topic = consumer.topic.name
           kafka_messages = event[:kafka_batch].messages
-          info "#{kafka_messages.count} messages on #{topic} topic delegated to #{consumer.class}"
+          info(
+            <<~MSG.chomp.tr("\n", ' ')
+              #{kafka_messages.count} messages
+              on #{topic} topic
+              delegated to #{consumer.class}
+            MSG
+          )
         end
 
         # Logs details about incoming message and with which consumer we will consume it
@@ -34,18 +40,23 @@ module Karafka
           info "1 message on #{topic} topic delegated to #{consumer.class}"
         end
 
-        # Logs details about each received message value parsing
+        # Logs details about each received message value deserialization
         # @param event [Dry::Events::Event] event details including payload
-        def on_params_params_parse(event)
+        def on_params_params_deserialize(event)
           # Keep in mind, that a caller here is a param object not a controller,
           # so it returns a topic as a string, not a routing topic
-          debug "Params parsing for #{event[:caller].topic} topic successful in #{event[:time]} ms"
+          debug(
+            <<~MSG.chomp.tr("\n", ' ')
+              Params deserialization for #{event[:caller].topic} topic
+              successful in #{event[:time]} ms
+            MSG
+          )
         end
 
-        # Logs unsuccessful parsing attempts of incoming data
+        # Logs unsuccessful deserialization attempts of incoming data
         # @param event [Dry::Events::Event] event details including payload
-        def on_params_params_parse_error(event)
-          error "Params parsing error for #{event[:caller].topic} topic: #{event[:error]}"
+        def on_params_params_deserialize_error(event)
+          error "Params deserialization error for #{event[:caller].topic} topic: #{event[:error]}"
         end
 
         # Logs errors that occured in a listener fetch loop