+++ lib/karafka/schemas/config.rb
@@ -5,7 +5,7 @@ module Karafka
   module Schemas
     # Regexp for validating format of groups and topics
     # @note It is not nested inside of the schema, as it is used by couple schemas
-    TOPIC_REGEXP = /\A(\w|\-|\.)+\z/
+    TOPIC_REGEXP = /\A(\w|\-|\.)+\z/.freeze
 
     # Schema with validation rules for Karafka configuration details
     # @note There are many more configuration options inside of the