+++ spec/lib/karafka/schemas/consumer_group_spec.rb
@@ -9,7 +9,7 @@ RSpec.describe Karafka::Schemas::ConsumerGroup do
         name: 'name',
         backend: :inline,
         consumer: Class.new,
-        parser: Class.new,
+        deserializer: Class.new,
         max_bytes_per_partition: 1,
         start_from_beginning: true,
         batch_consuming: true,